refactor: split inline test modules from 9 more userspace-dp files
Second batch of the pure-code-motion test-split modularity work (after
#4779 poll_stages and #4780's eight-file batch). Relocate the tail
`#[cfg(test)] mod tests { ... }` block out of nine more modules into sibling
`<name>_tests.rs` files, loaded via `#[cfg(test)] #[path = "..."] mod
tests;`. Part of the #4404/#4408/#4421 modularity backlog.

  io_uring_write.rs   881 -> ~418   (462 test lines out)
  afxdp/icmp.rs       1014 -> ~598  (415 out)
  poll_descriptor/cookie_reply.rs 509 -> ~129 (379 out)
  screen/rate.rs      609 -> ~268   (340 out)
  afxdp/icmp_ratelimit.rs 703 -> ~374 (328 out)
  afxdp/wg/handshake.rs 651 -> ~337 (313 out)
  afxdp/forwarding/host_inbound.rs 817 -> ~536 (280 out)
  screen/syn_rate.rs  504 -> ~275   (228 out)
  fairness_eval/per_worker.rs 475 -> ~253 (221 out)

Pure code motion, verified per file:
- Relocated body is byte-identical: each sibling was populated by moving the
  `mod tests { ... }` inner body VERBATIM (not a dedent transform), then
  rustfmt'd (edition 2024, syntax-aware) purely to canonicalise indentation.
- Each parent's production code (everything before the test module) is
  byte-identical to origin/master; the only parent change is the three-line
  `mod tests {...}` -> `#[path] mod tests;` swap (parents are NOT reformatted).
- Test count preserved per file: 15/14/5/14/10/14/6/10/13 #[test]/
  #[tokio::test] before and after. Each module keeps `use super::*` and the
  `#[path]` child keeps its module parent, so `super::` paths are unchanged.

Validation: cargo build clean (all nine #[path] splits resolve); FULL
cargo test --release to follow before merge. No production behavior change.
Refs #4404 #4408 #4421.
