+++ userspace-dp/src/screen/rate.rs
@@ -265,345 +265,5 @@ impl TokenBucket {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// One whole nanosecond-second, for readable `now_ns` arithmetic in tests.
-    const NS: u64 = 1_000_000_000;
-
-    /// Within a single second the sliding window behaves like the old
-    /// fixed window: `threshold` events pass, the next one trips.
-    #[test]
-    fn within_window_allows_up_to_threshold() {
-        let mut rc = RateCounter::default();
-        for _ in 0..5 {
-            assert!(!rc.increment(100, 5), "first 5 events must be admitted");
-        }
-        assert!(rc.increment(100, 5), "6th event in the same second trips");
-    }
-
-    /// FAIL-ON-REVERT (#2937): a boundary double-burst must NOT admit 2x
-    /// the threshold. `threshold` events at the end of second N followed
-    /// by another `threshold` at the start of second N+1 straddle the
-    /// window boundary in a sub-millisecond interval. The fixed-window
-    /// reset admitted all 2N; the sliding window trips on the first event
-    /// of second N+1 because second N's tally still counts.
-    ///
-    /// Against the reverted (fixed-window) code this assertion is RED: the
-    /// fixed window resets `count` at the boundary so all 2N pass and
-    /// `tripped_after_boundary` stays 0.
-    #[test]
-    fn boundary_double_burst_is_bounded() {
-        const THRESHOLD: u32 = 100;
-        let mut rc = RateCounter::default();
-
-        // End of second N: consume the full budget. None trip.
-        let mut admitted_n = 0u32;
-        for _ in 0..THRESHOLD {
-            if !rc.increment(0, THRESHOLD) {
-                admitted_n += 1;
-            }
-        }
-        assert_eq!(admitted_n, THRESHOLD, "full budget admitted in second N");
-
-        // Start of second N+1 (boundary crossing): a second full burst.
-        // The trailing 1-second sum is prev(THRESHOLD) + current, so the
-        // very first event already exceeds the threshold.
-        let mut tripped_after_boundary = 0u32;
-        for _ in 0..THRESHOLD {
-            if rc.increment(1, THRESHOLD) {
-                tripped_after_boundary += 1;
-            }
-        }
-        assert_eq!(
-            tripped_after_boundary, THRESHOLD,
-            "every event in the post-boundary burst must trip (fixed \
-             window would admit all 2N and trip none)"
-        );
-    }
-
-    /// A two-second idle gap clears the previous bucket — a sender that
-    /// goes quiet and returns gets a fresh full budget (the carryover is
-    /// only one second deep, matching the sliding-window definition).
-    #[test]
-    fn gap_clears_previous_bucket() {
-        const THRESHOLD: u32 = 10;
-        let mut rc = RateCounter::default();
-        for _ in 0..THRESHOLD {
-            assert!(!rc.increment(0, THRESHOLD));
-        }
-        // Jump two seconds ahead: second 1 was empty, so the trailing
-        // window no longer contains second 0's events.
-        let mut admitted = 0u32;
-        for _ in 0..THRESHOLD {
-            if !rc.increment(2, THRESHOLD) {
-                admitted += 1;
-            }
-        }
-        assert_eq!(admitted, THRESHOLD, "fresh budget after a full-window gap");
-    }
-
-    /// #3315 D7: `increment_and_classify` advances the window exactly ONCE and
-    /// classifies the same trailing sum against both thresholds. Each call
-    /// counts exactly one event, so the sum climbs 1,2,3,... and crosses the
-    /// alarm threshold before the (higher) attack threshold. A peek-then-
-    /// increment implementation that advanced the window twice would miscount;
-    /// this sequence pins the single-advance, dual-compare contract.
-    #[test]
-    fn increment_and_classify_single_advance_dual_threshold() {
-        const ATTACK: u32 = 5;
-        const ALARM: u32 = 2;
-        let mut rc = RateCounter::default();
-        // sum 1,2: under both (2 > 2 is false — boundary is ">").
-        assert_eq!(rc.increment_and_classify(0, ATTACK, ALARM), (false, false));
-        assert_eq!(rc.increment_and_classify(0, ATTACK, ALARM), (false, false));
-        // sum 3: over alarm (3 > 2), under attack (3 > 5 false).
-        assert_eq!(rc.increment_and_classify(0, ATTACK, ALARM), (false, true));
-        // sum 4, 5: still over alarm, under/at attack (5 > 5 false).
-        assert_eq!(rc.increment_and_classify(0, ATTACK, ALARM), (false, true));
-        assert_eq!(rc.increment_and_classify(0, ATTACK, ALARM), (false, true));
-        // sum 6: over both.
-        assert_eq!(rc.increment_and_classify(0, ATTACK, ALARM), (true, true));
-    }
-
-    /// A sustained sender at exactly the threshold rate stays admitted:
-    /// the configured `N/sec` semantic is preserved across many seconds.
-    #[test]
-    fn sustained_at_threshold_is_admitted() {
-        const THRESHOLD: u32 = 50;
-        // Split the per-second budget half before and half after each
-        // boundary; the trailing-window sum is THRESHOLD/2 + THRESHOLD/2
-        // == THRESHOLD, which is admitted (the boundary case is "> not
-        // >=").
-        let mut rc = RateCounter::default();
-        let half = THRESHOLD / 2;
-        for sec in 0..20u64 {
-            for _ in 0..half {
-                assert!(
-                    !rc.increment(sec, THRESHOLD),
-                    "sustained sender at threshold must stay admitted at sec {sec}"
-                );
-            }
-        }
-    }
-
-    /// #3607 CONTRAST: `RateCounter` fed a FULL `threshold` events/second at
-    /// the second boundary over-throttles to ~0 after the first second — the
-    /// bug that motivates the token bucket. `sustained_at_threshold_is_admitted`
-    /// only passes because it feeds `threshold/2`; at the real threshold the
-    /// window saturates and rejects nearly everything.
-    #[test]
-    fn rate_counter_over_throttles_full_threshold() {
-        const THRESHOLD: u32 = 100;
-        let mut rc = RateCounter::default();
-        let mut admitted = 0u32;
-        for sec in 0..10u64 {
-            for _ in 0..THRESHOLD {
-                if !rc.increment(sec, THRESHOLD) {
-                    admitted += 1;
-                }
-            }
-        }
-        // Only the very first second's budget gets through; every later second
-        // stays saturated by the prior second's count-all tally.
-        assert!(
-            admitted <= THRESHOLD,
-            "RateCounter admits only ~one window of a sustained at-threshold \
-             stream (got {admitted} over 10 windows of {THRESHOLD})"
-        );
-    }
-
-    /// #3607 (M09): the token bucket admits a sustained sender parked at
-    /// EXACTLY `threshold` events/second — the case `RateCounter` throttles to
-    /// ~0. Events are evenly spaced at the real monotonic rate (`1s/threshold`
-    /// apart), so each event accrues exactly one token before it is spent.
-    #[test]
-    fn token_bucket_sustained_at_threshold_admits() {
-        const THRESHOLD: u32 = 1000;
-        let interval_ns = NS / THRESHOLD as u64; // 1_000_000 ns == 1ms
-        let mut tb = TokenBucket::default();
-        let mut now = 5 * NS; // arbitrary non-zero monotonic base
-        let mut admitted = 0u32;
-        let total = THRESHOLD * 20; // 20 seconds of sustained at-threshold load
-        for _ in 0..total {
-            if !tb.admit_is_over(now, THRESHOLD) {
-                admitted += 1;
-            }
-            now += interval_ns;
-        }
-        // Continuous refill keeps the bucket topped up: essentially everything
-        // is admitted (>= 99%), versus RateCounter which would drop ~95%.
-        assert!(
-            admitted as f64 >= 0.99 * total as f64,
-            "sustained at-threshold must stay admitted: {admitted}/{total}"
-        );
-    }
-
-    /// Low-threshold sanity: `threshold == 1` sustained at 1 pps admits every
-    /// second (the `T=1` roughness that killed the weighted-window option).
-    #[test]
-    fn token_bucket_low_threshold_sustained_admits() {
-        let mut tb = TokenBucket::default();
-        let mut now = NS;
-        for sec in 0..30u64 {
-            assert!(
-                !tb.admit_is_over(now, 1),
-                "1 pps at threshold 1 must be admitted every second (sec {sec})"
-            );
-            now += NS;
-        }
-    }
-
-    /// A fresh bucket cold-starts FULL: the first `threshold` events at one
-    /// instant are admitted, matching `RateCounter`'s first window.
-    #[test]
-    fn token_bucket_cold_start_admits_first_threshold() {
-        const THRESHOLD: u32 = 50;
-        let mut tb = TokenBucket::default();
-        let now = 7 * NS;
-        for i in 0..THRESHOLD {
-            assert!(
-                !tb.admit_is_over(now, THRESHOLD),
-                "cold-start event {i} within capacity must be admitted"
-            );
-        }
-        assert!(
-            tb.admit_is_over(now, THRESHOLD),
-            "the (threshold+1)th event at the same instant is over capacity"
-        );
-    }
-
-    /// #2937 RETAINED: a burst ABOVE capacity at a single instant (and a
-    /// sub-millisecond straddle) is bounded to `threshold` — the bucket never
-    /// hands out 2x the budget across a boundary the way a fixed window would.
-    #[test]
-    fn token_bucket_burst_above_capacity_is_bounded() {
-        const THRESHOLD: u32 = 100;
-        let mut tb = TokenBucket::default();
-        let now = 3 * NS;
-        let mut admitted = 0u32;
-        for _ in 0..(THRESHOLD * 3) {
-            if !tb.admit_is_over(now, THRESHOLD) {
-                admitted += 1;
-            }
-        }
-        assert_eq!(admitted, THRESHOLD, "instantaneous burst bounded to capacity");
-        // Sub-millisecond straddle 100us later: negligible refill, still bounded.
-        let later = now + 100_000; // 100us
-        let mut admitted2 = 0u32;
-        for _ in 0..THRESHOLD {
-            if !tb.admit_is_over(later, THRESHOLD) {
-                admitted2 += 1;
-            }
-        }
-        assert!(
-            admitted2 <= 1,
-            "a sub-ms micro-burst must not refill a second budget (got {admitted2})"
-        );
-    }
-
-    /// After an over-limit burst, dropping back to <= threshold recovers WITHOUT
-    /// needing a fully idle second (the RateCounter suppress-until-idle defect).
-    #[test]
-    fn token_bucket_recovers_without_idle_second() {
-        const THRESHOLD: u32 = 10;
-        let mut tb = TokenBucket::default();
-        let mut now = NS;
-        // Drain the bucket with an instantaneous 5x burst.
-        for _ in 0..(THRESHOLD * 5) {
-            tb.admit_is_over(now, THRESHOLD);
-        }
-        // 200ms later at <= threshold pacing: 200ms accrues 0.2*threshold == 2
-        // tokens, so at least the next couple of events are admitted again
-        // (no full idle second required).
-        now += NS / 5;
-        assert!(
-            !tb.admit_is_over(now, THRESHOLD),
-            "partial refill after 200ms must admit again without an idle second"
-        );
-    }
-
-    /// An out-of-order (`now_ns < last_refill_ns`) clock sample must not panic
-    /// and must not over-credit — `saturating_sub` yields 0 elapsed.
-    #[test]
-    fn token_bucket_clock_underflow_no_panic() {
-        const THRESHOLD: u32 = 8;
-        let mut tb = TokenBucket::default();
-        // Initialise at a high base.
-        assert!(!tb.admit_is_over(100 * NS, THRESHOLD));
-        // A sample far in the past: no refill, no panic; the surviving tokens
-        // still admit up to capacity-1 more.
-        for _ in 0..(THRESHOLD - 1) {
-            assert!(!tb.admit_is_over(1, THRESHOLD));
-        }
-        assert!(
-            tb.admit_is_over(1, THRESHOLD),
-            "no spurious refill from a backwards clock"
-        );
-    }
-
-    /// RED-ON-REVERT (Copilot #4321): a backwards clock step must not rewind
-    /// the refill high-water mark, or a later forward step over-credits the
-    /// bucket and the rate limit is weakened by a clock glitch. Drain at
-    /// `base`, step BACKWARDS 5s (refills 0 — must NOT move `last_refill_ns` to
-    /// the dip), then step forward only 1us past `base`. With the monotonic
-    /// high-water mark the elapsed since the true last refill is ~1us (~0
-    /// refill) so the bucket stays limited. On the buggy code `last_refill_ns`
-    /// dipped to `base-5s`, so the forward call sees ~5s elapsed (capped at 1s
-    /// == a FULL capacity refill) and admits a fresh budget — RED.
-    #[test]
-    fn token_bucket_backwards_clock_does_not_over_credit() {
-        const THRESHOLD: u32 = 10;
-        let mut tb = TokenBucket::default();
-        let base = 10 * NS;
-        // Drain the bucket at `base`.
-        for _ in 0..(THRESHOLD * 2) {
-            tb.admit_is_over(base, THRESHOLD);
-        }
-        assert!(tb.admit_is_over(base, THRESHOLD), "bucket drained at base");
-        // Backwards clock step (5s earlier): 0 refill AND no rewind of the mark.
-        assert!(
-            tb.admit_is_over(base - 5 * NS, THRESHOLD),
-            "backwards step refills nothing — still limited"
-        );
-        // Forward again, only 1us after `base`: ~0 elapsed since the true last
-        // refill, so the bucket must STILL be limited (0 admitted).
-        let mut admitted = 0u32;
-        for _ in 0..THRESHOLD {
-            if !tb.admit_is_over(base + 1_000, THRESHOLD) {
-                admitted += 1;
-            }
-        }
-        assert_eq!(
-            admitted, 0,
-            "a backwards-then-forward clock step must not refill the bucket \
-             (monotonic high-water mark): got {admitted} admitted"
-        );
-    }
-
-    /// A multi-second idle gap at a very high threshold does not overflow the
-    /// `elapsed * threshold` fixed-point multiply (elapsed capped at 1s), and
-    /// the bucket refills to exactly capacity.
-    #[test]
-    fn token_bucket_high_threshold_gap_no_overflow() {
-        const THRESHOLD: u32 = 1_000_000;
-        let mut tb = TokenBucket::default();
-        assert!(!tb.admit_is_over(NS, THRESHOLD)); // cold start full
-        // Drain a bit.
-        for _ in 0..100 {
-            tb.admit_is_over(NS, THRESHOLD);
-        }
-        // Jump 5 seconds ahead: elapsed caps at 1s, refills to full capacity.
-        let now = 6 * NS;
-        let mut admitted = 0u64;
-        for _ in 0..THRESHOLD {
-            if !tb.admit_is_over(now, THRESHOLD) {
-                admitted += 1;
-            }
-        }
-        assert_eq!(
-            admitted, THRESHOLD as u64,
-            "a >=1s gap refills to exactly capacity with no overflow"
-        );
-    }
-}
+#[path = "rate_tests.rs"]
+mod tests;
